+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Fixed a bug where the illuminants where not fully shown
+- The names of the GNOME color palette are now available
 
 ## [0.4.0] - 2022-10-20